% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundDown.R
\name{roundDown}
\alias{roundDown}
\title{Round down a negative number}
\usage{
roundDown(x, dig)
}
\arguments{
\item{x}{number to be rounded}

\item{dig}{number of decimal places to be rounded}
}
\description{
Rounds a negative number to the nearest, more negative
number to a specified decimal place.
}
\details{
\code{roundDown} is used to
round a negative Z-statistic critical value, which has (infinitely)
many significant digits, to the nearest \code{dig} decimal place
which is more extreme
than the critical value itself.  This is done to preserve the
desired type I error level.
}
\examples{
roundDown(-1.95456, 2)
# [1] -1.96
}
