% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor2x2.R
\name{cor2x2}
\alias{cor2x2}
\title{Hazard ratios and correlations for the 2x2 statistics}
\usage{
cor2x2(time, event, indA, indB, covmat)
}
\arguments{
\item{time}{follow-up times}

\item{event}{event indicators (0/1)}

\item{indA}{treatment A indicators (0/1)}

\item{indB}{treatment B indicators (0/1)}

\item{covmat}{matrix of covariates; one row per subject.  NOTE!! Factor variables must
use 0/1 indicator variables}
}
\value{
\item{loghrA }{overall A log hazard ratio}

\item{seA }{standard error of the overall A log hazard ratio}

\item{hrA }{overall A hazard ratio}

\item{ciA }{95\% confidence interval for overall A hazard ratio}

\item{pvalA }{two-sided p-value for overall A hazard ratio}

\item{loghra }{simple A log hazard ratio}

\item{sea }{standard error of the simple A log hazard ratio}

\item{hra }{simple A hazard ratio}

\item{cia }{95\% confidence interval for simple A hazard ratio}

\item{pvala }{two-sided p-value for simple A hazard ratio}

\item{loghrab }{simple AB log hazard ratio}

\item{seab}{standard error of the simple AB log hazard ratio}

\item{hrab }{simple AB hazard ratio}

\item{ciab}{95\% confidence interval for simple AB hazard ratio}

\item{pvalab }{two-sided p-value for simple AB hazard ratio}

\item{corAa }{correlation between the overall A and simple A test statistics}

\item{corAab }{correlation between the overall A and simple AB test statistics}

\item{coraab }{correlation between the simple A and simple AB test statistics}
}
\description{
Computes  the hazard ratios, confidence intervals, p-values, and correlations
for the overall A, simple A, and simple AB test statistics.
}
\details{
This function computes
(i) correlation between the overall A test and the simple A test
(ii) correlation between the overall A test and the simple AB test
(iii) correaltion between the simple A and simple AB test.
The correlation estimates are derived in
Lin, Gong, Gallo, et al. (Biometrics 2016).
}
\examples{
\dontrun{
# Not run by CRAN.  Requires less than 10 seconds on a personal computer.
# First load the simulated data variables. The "simdata" file is
# a 4600-by-9 matrix which is loaded with the factorial2x2 package.
time <- simdata[, "time"]
event <- simdata[, "event"]
indA <- simdata[, "indA"]
indB <- simdata[, "indB"]
covmat <- simdata[, 6:10]

cor2x2(time, event, indA, indB, covmat)
# $loghrA
# [1] -0.1170805

# $seA
# [1] 0.06258749

# hrA
# [1] 0.8895135

# ciA
# [1] 0.786823 1.005607

# pvalA
# [1] 0.06139083

# $loghra
# [1] -0.2112048

# $sea
# [1] 0.08655462

# hra
# [1] 0.8096082

# cia
# [1] 0.6832791 0.9592939

# pvala
# [1] 0.01468184

# hrab
# [1] 0.7583061

# $loghrab
# [1] -0.2766681

# $seab
# [1] 0.08738966

# ciab
# [1] 0.6389355 0.8999785

# pvalab
# [1] 0.001545967

# corAa
# [1] 0.7274961

# corAab
# [1] 0.7164075

# coraab
# [1] 0.4572905
}
}
\references{
Lin, D.Y., Glong , J., Gallo, P., Bunn, P.H., Couper, D.
Simultaneous inference on treatment effects in survival studies
with factorial designs. Biometrics, 2016; 72: 1078-1085.
}
