% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fac2x2analyze.R
\name{fac2x2analyze}
\alias{fac2x2analyze}
\title{Significance testing for the 2/3-1/3, 1/3-1/3-1/3, 1/2-1/2 procedures}
\usage{
fac2x2analyze(time, event, indA, indB, covmat, alpha, dig = 2,
  niter = 5)
}
\arguments{
\item{time}{follow-up times}

\item{event}{event indicators (0/1)}

\item{indA}{treatment A indicators (0/1)}

\item{indB}{treatment B indicators (0/1)}

\item{covmat}{covariate matrix, must be non-NULL.  Factor variables MUST use 0/1 dummy variables}

\item{alpha}{two-sided familywise significance level}

\item{dig}{number of decimal places to which we \code{\link{roundDown}} the critical value}

\item{niter}{number of interations passed to \code{crit2x2} function call}
}
\value{
\item{loghrA }{overall A log hazard ratio}

\item{seA }{standard error of the overall A log hazard ratio}

\item{ZstatA }{Z-statistic for the overall A log hazard ratio}

\item{pvalA }{two-sided p-value for the overall hazard ratio}

\item{hrA }{overall A hazard ratio}

\item{ciA }{95\% confidence interval for the overall A hazard ratio}

\item{loghra }{simple A log hazard ratio}

\item{sea }{standard error of the simple A log hazard ratio}

\item{Zstata }{Z-statistic for the simple A log hazard ratio}

\item{pvala }{two-sided p-value for the simple A hazard ratio}

\item{hra }{simple A hazard ratio}

\item{cia }{95\% confidence interval for the simple A hazard ratio}

\item{loghrab }{simple AB log hazard ratio}

\item{seab}{standard error of the simple AB log hazard ratio}

\item{Zstatab }{Z-statistic for the simple AB log hazard ratio}

\item{pvalab }{two-sided p-value for the simple AB hazard ratio}

\item{hrab }{simple AB hazard ratio}

\item{ciab}{95\% confidence interval for the simple AB hazard ratio}

\item{crit23A }{2/3-1/3 procedure's critical value for the overall A statistic}

\item{sig23A }{2/3-1/3 procedure's p-value rejection criterion for the overall A null hypothesis}

\item{crit23ab}{2/3-1/3 procedure's critical value for the simple AB statistic}

\item{sig23ab }{2/3-1/3 procedure's p-value rejection criterion for the simple A null hypothesis}

\item{result23 }{2/3-1/3 procedure's accept/reject decisions for the overall A and simple A null hypotheses results}

\item{crit13 }{1/3-1/3-1/3 procedure's critical value for all three test statistics}

\item{sig13 }{1/3-1/3-1/3 procedure's p-value rejection criterion for the overall A, simple A, and simple AB null hypotheses}

\item{result13 }{1/3-1/3-1/3 procedure's accept/reject decisions for the overall A, simple A, and simple AB null hypotheses results}

\item{crit12 }{1/2-1/2 procedure's critical value for the simple A and AB statistics}

\item{sig12 }{1/2-1/2 procedure's p-value rejection criterion for the simple A and simple AB null hypotheses}

\item{result12 }{1/2-1/2 procedure's accept/reject decisions for the simple A and simple AB null hypotheses results}

\item{corAa }{correlation between the overall A and simple A Wald statistics}

\item{corAab }{correlation between the overall A and simple AB Wald statistics}

\item{coraab }{correlation between the simple A and simple AB Wald statistics}
}
\description{
Performs significance testing for the 2/3-1/3, 1/3-1/3-1/3, 1/2-1/2 procedures.
Also reports the hazard ratios, 95\% confidence intervals, p-values,
nominal significance levels, and correlations for the overall A,
simple A, and simple AB test statistics.
}
\examples{
\dontrun{
 # Not run by CRAN.  Requires less than 15 seconds on a personal computer.
 # First load the simulated data variables. The "simdata" file is
 # a 4600-by-9 matrix which is loaded with the factorial2x2 package.
 time <- simdata[, "time"]
 event <- simdata[, "event"]
 indA <- simdata[, "indA"]
 indB <- simdata[, "indB"]
 covmat <- simdata[, 6:10]
 fac2x2analyze(time, event, indA, indB, covmat, alpha = 0.05, niter = 5)
# $loghrA
# [1] -0.1170805

# $seA
# [1] 0.06258749

# $ZstatA
# [1] -1.87067

# $hrA
# [1] 0.8895135

# $pvalA
# [1] 0.06139083

# $ciA
# [1] 0.786823 1.005607

# $loghra
# [1] -0.2112048

# $sea
# [1] 0.08655462

# $Zstata
# [1] -2.440133

# $pvala
# [1] 0.01468184

# $hra
# [1] 0.8096082

# $cia
# [1] 0.6832791 0.9592939

# $loghrab
# [1] -0.2766681

# $seab
# [1] 0.08738966

# $Zstatab
# [1] -3.165914

# $pvalab
# [1] 0.001545967

# $hrab
# [1] 0.7583061

# $ciab
# [1] 0.6389355 0.8999785

# $pvalab
# [1] 0.001545967

# $sig23A
# [1] 0.03333294

# $sig23ab
# [1] 0.02560439

# $result23
# [1] "accept overall A" "reject simple AB"

# $sig13
# [1] 0.02091308

# $result13
# [1] "accept overall A" "reject simple A"  "reject simple AB"

# $sig12
# [1] 0.02665044

# $result12
# [1] "reject simple A"  "reject simple AB"

# $corAa
# [1] 0.7274961

# $corAab
# [1] 0.7164075

# $coraab
# [1] 0.4572905
}
}
\references{
Leifer, E.S., Troendle, J.F., Kolecki, A., Follmann, D.
Joint testing of overall and simple effect for the two-by-two factorial design. (2019). Submitted.
}
\author{
Eric Leifer, James Troendle
}
